+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+- If a stop is marked as a timepoint in any trips, treat it as a timepoint in all trips for sorting purposes.
+
 ## [2.10.11] - 2025-04-17
 
 ### Updated